:sparkles: Handle proper URL ?client_id=
