+++ cmd/experiments/web-ui/handlers.go
@@ -1,0 +1,197 @@
+package main
+
+import (
+	"context"
+	"fmt"
+	"html/template"
+	"net/http"
+	"strings"
+	"time"
+
+	"github.com/google/uuid"
+)
+
+// RegisterHandlers sets up all HTTP handlers for the server
+func RegisterHandlers(server *Server, tmpl *template.Template) {
+	// Serve index page
+	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
+		// Check if there's a client ID in the URL
+		clientID := r.URL.Query().Get("client_id")
+
+		// Verify client exists if ID provided
+		if clientID != "" {
+			server.clientsMux.RLock()
+			_, exists := server.clients[clientID]
+			server.clientsMux.RUnlock()
+			if !exists {
+				// Client doesn't exist, redirect to root
+				server.logger.Info().Str("client_id", clientID).Msg("Client not found, redirecting to root")
+				http.Redirect(w, r, "/", http.StatusTemporaryRedirect)
+				return
+			}
+		}
+
+		data := TemplateData{
+			ClientID: clientID,
+		}
+
+		if err := tmpl.ExecuteTemplate(w, "index.html", data); err != nil {
+			server.logger.Error().Err(err).Msg("Failed to render index page")
+			http.Error(w, err.Error(), http.StatusInternalServerError)
+		}
+	})
+
+	// Handle SSE events endpoint
+	http.HandleFunc("/events", func(w http.ResponseWriter, r *http.Request) {
+		// Get client ID from query params
+		clientID := r.URL.Query().Get("client_id")
+		server.logger.Info().Str("client_id", clientID).Msg("Handling SSE events")
+		if clientID == "" {
+			server.logger.Error().Msg("No client ID provided for events")
+			http.Error(w, "No client ID provided", http.StatusBadRequest)
+			return
+		}
+
+		// Set SSE headers
+		w.Header().Set("Content-Type", "text/event-stream")
+		w.Header().Set("Cache-Control", "no-cache")
+		w.Header().Set("Connection", "keep-alive")
+		w.Header().Set("Access-Control-Allow-Origin", "*")
+
+		flusher, ok := w.(http.Flusher)
+		if !ok {
+			server.logger.Error().Msg("Streaming unsupported")
+			http.Error(w, "Streaming unsupported!", http.StatusInternalServerError)
+			return
+		}
+
+		// Get existing client or create new one
+		server.clientsMux.RLock()
+		client, exists := server.clients[clientID]
+		server.clientsMux.RUnlock()
+
+		if !exists {
+			server.logger.Info().Str("client_id", clientID).Msg("Creating new client")
+			client = NewSSEClient(clientID)
+			server.RegisterClient(client)
+		}
+		defer server.UnregisterClient(client.ID)
+
+		server.logger.Info().Str("client_id", client.ID).Msg("SSE connection established")
+
+		// Write events to client with timeout to prevent stuck connections
+		for {
+			select {
+			case <-r.Context().Done():
+				server.logger.Info().Str("client_id", client.ID).Msg("SSE connection closed by client")
+				return
+			case msg, ok := <-client.MessageChan:
+				if !ok {
+					server.logger.Info().Str("client_id", client.ID).Msg("SSE message channel closed")
+					return
+				}
+				// Handle multiline messages by prefacing each line with data:
+				lines := strings.Split(msg, "\n")
+				fmt.Fprintf(w, "event: message\n")
+				for _, line := range lines {
+					fmt.Fprintf(w, "data: %s\n", line)
+				}
+				fmt.Fprintf(w, "\n")
+				flusher.Flush()
+			case <-time.After(30 * time.Second):
+				// Send heartbeat to keep connection alive
+				fmt.Fprintf(w, "event: heartbeat\ndata: ping\n\n")
+				flusher.Flush()
+				server.logger.Debug().Str("client_id", client.ID).Msg("Sent heartbeat")
+			}
+		}
+	})
+
+	// Handle start endpoint - creates and starts a new step
+	http.HandleFunc("/start", func(w http.ResponseWriter, r *http.Request) {
+		// Create new client ID
+		clientID := uuid.New().String()
+		server.logger.Info().Str("client_id", clientID).Msg("Starting new chat")
+
+		// Create new client
+		client := NewSSEClient(clientID)
+		server.RegisterClient(client)
+
+		if err := server.CreateStep(clientID); err != nil {
+			server.logger.Error().Err(err).Str("client_id", clientID).Msg("Failed to create step")
+			http.Error(w, fmt.Sprintf("Error creating step: %v", err), http.StatusInternalServerError)
+			return
+		}
+
+		// XXX we need to properly handle the context here, at a client level
+		err := server.router.RunHandlers(context.Background())
+		if err != nil {
+			server.logger.Error().Err(err).Str("client_id", clientID).Msg("Failed to run handlers")
+			http.Error(w, fmt.Sprintf("Error running handlers: %v", err), http.StatusInternalServerError)
+			return
+		}
+
+		if err := server.StartStep(clientID); err != nil {
+			server.logger.Error().Err(err).Str("client_id", clientID).Msg("Failed to start step")
+			http.Error(w, fmt.Sprintf("Error starting step: %v", err), http.StatusInternalServerError)
+			return
+		}
+
+		// Return the chat container template with the new client ID
+		data := TemplateData{
+			ClientID: clientID,
+		}
+		w.Header().Set("HX-Push-Url", fmt.Sprintf("/?client_id=%s", clientID))
+		if err := tmpl.ExecuteTemplate(w, "chat-container", data); err != nil {
+			server.logger.Error().Err(err).Msg("Failed to render chat container")
+			http.Error(w, err.Error(), http.StatusInternalServerError)
+			return
+		}
+	})
+
+	// Handle chat endpoint - processes chat messages
+	http.HandleFunc("/chat", func(w http.ResponseWriter, r *http.Request) {
+		if r.Method != http.MethodPost {
+			http.Error(w, "Method not allowed", http.StatusMethodNotAllowed)
+			return
+		}
+
+		clientID := r.FormValue("client_id")
+		message := r.FormValue("message")
+
+		if clientID == "" || message == "" {
+			http.Error(w, "Missing client_id or message", http.StatusBadRequest)
+			return
+		}
+
+		server.logger.Info().
+			Str("client_id", clientID).
+			Str("message", message).
+			Msg("Received chat message")
+
+		// Create new step for this message
+		if err := server.CreateStep(clientID); err != nil {
+			server.logger.Error().Err(err).Str("client_id", clientID).Msg("Failed to create step")
+			http.Error(w, fmt.Sprintf("Error creating step: %v", err), http.StatusInternalServerError)
+			return
+		}
+
+		// Run handlers
+		err := server.router.RunHandlers(context.Background())
+		if err != nil {
+			server.logger.Error().Err(err).Str("client_id", clientID).Msg("Failed to run handlers")
+			http.Error(w, fmt.Sprintf("Error running handlers: %v", err), http.StatusInternalServerError)
+			return
+		}
+
+		// Start step
+		if err := server.StartStep(clientID); err != nil {
+			server.logger.Error().Err(err).Str("client_id", clientID).Msg("Failed to start step")
+			http.Error(w, fmt.Sprintf("Error starting step: %v", err), http.StatusInternalServerError)
+			return
+		}
+
+		// Return empty response as events will be sent via SSE
+		w.WriteHeader(http.StatusOK)
+	})
+}
